feat: Renamed project-details.tsx file name
