+++ apps/web/components/pages/project/project-information/project-details.tsx
